[CAKER-2] Enable macOS target compatibility
- splitting toAttributes implementation in FontFace into AppKit/UIKit variants
- adjusted package platforms
